pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "agent-aibom"
version = "1.0.0"
description = "Agentic AI Bill of Materials — discover, graph, score, and export agent inventories"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "Agent AIBOM Contributors" }]
keywords = ["aibom", "ai", "agents", "security", "governance", "bom", "supply-chain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "pydantic>=2.7,<3",
    "typer>=0.12,<1",
    "rich>=13.7,<14",
    "networkx>=3.3,<4",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pyyaml>=6.0,<7",
    "jinja2>=3.1,<4",
    "python-frontmatter>=1.1,<2",
    "aiofiles>=24.1,<25",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.20,<6"]
grafana = ["grafana-client>=4.0,<5"]
dev = [
    "pytest>=8.2,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.5,<1",
    "mypy>=1.10,<2",
]

[project.scripts]
agent-aibom = "agent_aibom.cli:app"

[project.urls]
Homepage = "https://github.com/agent-aibom/agent-aibom"
Documentation = "https://agent-aibom.github.io"
Repository = "https://github.com/agent-aibom/agent-aibom"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
